requirements.txt
Django==1.8.8
psycopg2==2.6.1
Fabric==1.10.2
argparse==1.2.1
django-social-auth==0.7.28
ecdsa==0.13
freeze==1.0.9
httplib2==0.9.2
oauth2==1.9.0.post1
paramiko==1.16.0
pycrypto==2.6.1
python-openid==2.2.5
six==1.10.0
wsgiref==0.1.2
pyenchant==1.6.6
pytz==2015.7
python-dateutil==2.4.2

